+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Markdown Comments panel is now hidden by default. Enable with `workspaceShortcuts.markdownComments.panelEnabled` setting.
+
 ## [3.0.0] - 2026-01-11
 
 ### Added